Added Game Over Condition and restart
The snake will now detect collisions with the walls as well as itself. The food will also detect collisions with the snake and spawn in appropriate spots (not inside the snake. You can restart the game. You can hit game overs when the snake collides with anything. Need to fix the game speed because it scales horribly. We're almost done